requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
aiokafka
redis
pydantic[email]
pydantic-settings
python-dotenv
alembic
asyncpg
kafka-python

# Testing
pytest
pytest-asyncio
httpx 
